requirements.txt
Django>=4.2
plotly>=5.20.0
reportlab>=4.0.8
XlsxWriter>=3.2.0
supabase[all]>=2.0.3
python-dotenv>=1.0.1
requests>=2.31.0
gunicorn>=21.2.0